+++ dist-deno/schemas/array.ts
@@ -1,0 +1,12 @@
+import { ArraySchema, NullableOptions, OptionsForArray } from "../schemaClasses/ArraySchema.ts";
+export function array<T>(options: OptionsForArray<T> & NullableOptions): ArraySchema<T, null>;
+export function array<T>(options: OptionsForArray<T>): ArraySchema<T>;
+export function array<T>(): ArraySchema<T>;
+/**
+ * create schema
+ * @param options Options
+ * @returns schema
+ */
+export function array<T>(options: OptionsForArray<T> = {}): ArraySchema<T> {
+    return new ArraySchema<T>(options);
+}