+++ src/appliers/number/minValue.ts
@@ -39,7 +39,7 @@ export function applyTo(values: Values, options: Options, keyStack: Key[]): valu
 		return false;
 	}
 
-	ValueSchemaError.raise(CAUSE.MIN_VALUE, values, keyStack);
+	return ValueSchemaError.raise(CAUSE.MIN_VALUE, values, keyStack);
 }
 
 /**